backend/requirements.txt
albumentations==2.0.8
amqp==5.3.1
asgiref==3.11.0
billiard==4.2.4
celery==5.6.0
certifi==2025.11.12
channels==4.3.2
channels-redis==4.3.0
charset-normalizer==3.4.4
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
contourpy==1.3.3
<<<<<<< Updated upstream
coverage==7.12.0
=======
coverage==7.11.0    
>>>>>>> Stashed changes
cycler==0.12.1
Django==5.2.9
django-cors-headers==4.9.0
django-filter==25.2
django-redis==6.0.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.7
gunicorn==23.0.0
et_xmlfile==2.0.0
filelock==3.20.0
fonttools==4.60.1
fsspec==2025.10.0
huggingface-hub==0.36.0
idna==3.11
imageio==2.37.2
inflection==0.5.1
iniconfig==2.3.0
intel-openmp==2021.4.0
Jinja2==3.1.6
joblib==1.5.2
kiwisolver==1.4.9
kombu==5.6.1
lazy_loader==0.4
MarkupSafe==3.0.3
matplotlib==3.10.7
mkl==2021.4.0
mpmath==1.3.0
msgpack==1.1.2
networkx==3.6
numpy==2.1.3
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
Pillow==12.0.0
pluggy==1.6.0
prompt_toolkit==3.0.52
psutil==7.1.3
psycopg2-binary==2.9.11
py-cpuinfo==9.0.0
PyJWT==2.10.1
pyparsing==3.2.5
pytest==9.0.1
pytest-cov==7.0.0
pytest-django==4.11.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
qudida==0.0.4
redis==7.1.0
reportlab==4.0.4
requests==2.32.5
safetensors==0.7.0
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.16.2
seaborn==0.13.2
sendgrid==6.11.0
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
sympy==1.13.2
tbb==2021.13.1
threadpoolctl==3.6.0
tifffile==2025.10.16
timm==0.9.12
torch==2.5.1
torchvision==0.20.1
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.2
ultralytics==8.3.234
ultralytics-thop==2.0.17
uritemplate==4.2.0
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.14
XlsxWriter==3.1.9
whitenoise==6.8.2